requirements.txt
django

djangorestframework
django-cors-headers
markdown
django-filter

pytest
pytest-django
pytest-cov
mixer
snapshottest

